fish-classifier/requirements-api.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.3.0
torch>=2.2.0
torchvision>=0.17.0
python-dotenv>=1.0.0
